more clean + modular + new functions
