+++ bot/modes/__init__.py
@@ -1,0 +1,5 @@
+from bot.modes.mode import Mode
+from bot.modes.grind import Grind
+from bot.modes.exp_grind import ExperienceGrind
+from bot.modes.gold_grind import GoldGrind
+from bot.modes.charge_kindred import ChargeKindred
