+++ bot/modes/gold_grind.py
@@ -1,0 +1,8 @@
+from bot.modes import Grind
+
+class GoldGrind(Grind):
+
+    def main_loop(self):
+        
+        self.progress = self.character_details.get("dailyGold")/self.get_gold_cap()
+        super().main_loop()
