--- data/bosses.json
@@ -1,4 +1,0 @@
-{
-    "amboss": 19600,
-    "hybee": 41982
-}
