+++ templates/bosses/hybee.png
